CRUD: update finalizado

+++ trabalho_final_ls/geek_verso_shop/app/Http/Controllers/ProductController.php
@@ -1,4 +1,4 @@
-<?php 
+<?php
 
 namespace App\Http\Controllers;
 
@@ -33,19 +33,19 @@ class ProductController extends Controller
         //Salvar imagens na pasta storage/app/public
         $file = $request->file('product_file_name');
         $file_name = rand(10, 999999) . '-' . $file->getClientOriginalName();
-        $path = $request->file('product_file_name')->storeAs('image_uploads', $file_name);
+        $path = $file('product_file_name')->storeAs('image_uploads', $file_name);
 
         //Pegar os dados válidos do formulário
         $data = $request->validated();
         $data['product_file_name'] = $path;
         $data['user_id'] = Auth::user()->id;
-        
+
         //Salvar os dados no DB
         Product::create($data);
 
         //Redirecionar a página
         return redirect()->route('produtos.index')
-                        ->with('success', 'Produto adicionado com sucesso');
+            ->with('success', 'Produto adicionado com sucesso');
     }
 
     /**
@@ -59,7 +59,7 @@ class ProductController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Product $product)
+    public function edit(Product $produto)
     {
         //
     }
@@ -67,22 +67,50 @@ class ProductController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Product $product)
+    public function update(ProductRequest $request, Product $produto)
     {
-        //
+        //Pegar os dados válidos do formulário e transformar em um array
+        $data = $request->validated();
+
+        if ($request->hasFile('product_file_name')) {
+            //Deletar a imagem atual
+            Storage::disk('public')->delete($produto->product_file_name);
+
+            //Salvar a nova imagem na pasta storage/app/public
+            $file = $request->file('product_file_name');
+            $file_name = rand(10, 999999) . '-' . $file->getClientOriginalName();
+            $path = $file->storeAs('image_uploads', $file_name, 'public'); // <- o segredo
+            $data['product_file_name'] = $path;
+        }
+
+        //Atualizar os dados do DB
+        $data['product_name'] = $request->product_name;
+        $data['product_description'] = $request->product_description;
+        $data['product_quantity'] = $request->product_quantity;
+        $data['product_price'] = $request->product_price;
+        $data['product_rating'] = $request->product_rating;
+
+        $produto->user_id = Auth::user()->id;
+
+        //Fazer as alterações no DB
+        $produto->update($data);
+
+        //Redirecionar a página
+        return redirect()->route('produtos.index')
+            ->with('success', 'Produto editado com sucesso');
     }
 
     /**
      * Remove the specified resource from storage.
      */
     public function destroy(Product $produto)
-    {   
+    {
         //Deletar a imagem do produto a partir do disco onde está localizada a pasta public
-        Storage::disk('public')->delete($produto->product_file_name); 
+        Storage::disk('public')->delete($produto->product_file_name);
         $produto->delete();
 
         //Redirecionar a página
         return redirect()->route('produtos.index')
-                        ->with('success', 'Produto removido com sucesso');
+            ->with('success', 'Produto removido com sucesso');
     }
 }
